+++ src/common/js/configure.js
@@ -7,8 +7,9 @@ import _ from 'lodash';
 import { ReactiveVar } from 'meteor/reactive-var';
 
 let _conf = {};
+Modal._conf = new ReactiveVar( _conf );
 
-const _defaults = {
+Modal._defaults = {
     closeByBackdrop: true,
     contentClassesArray: null,
     verbosity: Modal.C.Verbose.CONFIGURE
@@ -23,12 +24,23 @@ const _defaults = {
  */
 Modal.configure = function( o ){
     if( o && _.isObject( o )){
-        _conf = _.merge( _defaults, _conf, o );
-        Modal._conf.set( _conf );
-        // be verbose if asked for
-        if( _conf.verbosity & Modal.C.Verbose.CONFIGURE ){
-            //console.log( 'pwix:modal configure() with', o, 'building', _conf );
-            console.log( 'pwix:modal configure() with', o );
+        // check that keys exist
+        let built_conf = {};
+        Object.keys( o ).forEach(( it ) => {
+            if( Object.keys( Modal._defaults ).includes( it )){
+                built_conf[it] = o[it];
+            } else {
+                console.warn( 'pwix:modal configure() ignore unmanaged key \''+it+'\'' );
+            }
+        });
+        if( Object.keys( built_conf ).length ){
+            _conf = _.merge( Modal._defaults, _conf, built_conf );
+            Modal._conf.set( _conf );
+            // be verbose if asked for
+            if( _conf.verbosity & Modal.C.Verbose.CONFIGURE ){
+                //console.log( 'pwix:modal configure() with', o, 'building', _conf );
+                console.log( 'pwix:modal configure() with', built_conf );
+            }
         }
     }
     // also acts as a getter
@@ -36,4 +48,4 @@ Modal.configure = function( o ){
 };
 
 _conf = _.merge( {}, Modal._defaults );
-Modal._conf = new ReactiveVar( _conf );
+Modal._conf .set( _conf );